requirements.txt
Pillow
python-dotenv
shortuuid
mongoengine
discord.py